requirements.txt
argcomplete==3.1.2
asgiref==3.7.2
boto3==1.28.68
botocore==1.31.68
certifi==2023.7.22
cfn-flip==1.3.0
charset-normalizer==3.3.0
click==8.1.7
Django==4.2.6
django-psdb-engine==1.0.6
django-storages==1.14.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
durationpy==0.5
gunicorn==21.2.0
hjson==3.1.0
idna==3.4
jmespath==1.0.1
kappa==0.6.0
MarkupSafe==2.1.3
packaging==23.2
placebo==0.9.0
psycopg2-binary==2.9.9
PyJWT==2.8.0
PyMySQL==1.1.0
python-dateutil==2.8.2
python-decouple==3.8
python-slugify==8.0.1
pytz==2023.3.post1
PyYAML==6.0.1
requests==2.28.2
s3transfer==0.7.0
six==1.16.0
sqlparse==0.4.4
stripe==7.0.0
text-unidecode==1.3
toml==0.10.2
tqdm==4.66.1
troposphere==4.4.1
typing_extensions==4.8.0
urllib3==1.26.18
Werkzeug==3.0.0
zappa==0.58.0
